+++ src/Sql/dbo/Tables/User.sql
@@ -32,7 +32,7 @@
     [ApiKey]                          VARCHAR (30)     NOT NULL,
     [ForcePasswordReset]              BIT              NOT NULL,
     [UsesKeyConnector]                BIT              NOT NULL,
-    [FailedLoginCount]                TINYINT          NOT NULL,
+    [FailedLoginCount]                INT              NOT NULL,
     [LastFailedLoginDate]             DATETIME2 (7)    NULL,
     CONSTRAINT [PK_User] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
